Move notifcationActionFunctions to shared/notifications and make it stateless
